+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-17 14:16:17
+
+### Changed Files
+- `src/services/hierarchy/invitationService.ts`
+
+### Statistics
+```
+ 1 file changed, 77 insertions(+), 33 deletions(-)
+```
+
 ## [Checkpoint] - 2025-12-17 13:49:04
 
 ### Changed Files